Update 1prep.sh
Trying to fix Maven install and download

+++ conf/docker-aio/1prep.sh
@@ -18,8 +18,12 @@ cd ../../
 cp -r scripts conf/docker-aio/testdata/
 cp doc/sphinx-guides/source/_static/util/createsequence.sql conf/docker-aio/testdata/doc/sphinx-guides/source/_static/util/
 
-echo "maven dependency prep"
-wget -q https://downloads.apache.org/maven/maven-3/3.8.6/binaries/apache-maven-3.8.6-bin.tar.gz -O dv/deps/apache-maven-3.8.6-bin.tar.gz
+
+if [ ! -e dv/deps/apache-maven-3.8.6-bin.tar.gz ]; then
+  echo "maven dependency prep"
+	wget -q https://downloads.apache.org/maven/maven-3/3.8.6/binaries/apache-maven-3.8.6-bin.tar.gz -O dv/deps/apache-maven-3.8.6-bin.tar.gz
+fi
+
 tar xfz dv/deps/apache-maven-3.8.6-bin.tar.gz
 mkdir maven
 mv apache-maven-3.8.6/* maven/
